expandable sections: show expanded when there's no javascript
Refactor the "show expandables and turn off expanded-state-realted UI" into a
mixin so we can use the same styles in both our print stylesheet as well as our
"no-js" layout styles.

If a browser doesn't have JS enabled, all the expandable sections are now
expanded and don't show any of the hover effects or other UI that indicate that
they're collapsible.